

%%%    Copyright (C) 2016-2026 Torsten Hothorn
%%%
%%%    This file is part of the 'libcoin' R add-on package.
%%%
%%%    'libcoin' is free software: you can redistribute it and/or modify
%%%    it under the terms of the GNU General Public License as published by
%%%    the Free Software Foundation, version 2.
%%%
%%%    'libcoin' is distributed in the hope that it will be useful,
%%%    but WITHOUT ANY WARRANTY; without even the implied warranty of
%%%    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
%%%    GNU General Public License for more details.
%%%
%%%    You should have received a copy of the GNU General Public License
%%%    along with 'libcoin'.  If not, see <http://www.gnu.org/licenses/>.
%%%
%%%
%%%    DO NOT EDIT THIS FILE
%%%
%%%    Edit 'libcoin.w' and run 'nuweb -r libcoin.w'

\name{ExpectationCovarianceStatistic}
\alias{ExpectationCovarianceStatistic}
\title{
  \pkg{libcoin} C API Example
}
\description{
  A working example of calling compiled code from the \pkg{libcoin} package.
}
\usage{
ExpectationCovarianceStatistic(x, y)
}
\arguments{
  \item{x}{A numeric matrix.}
  \item{y}{A numeric matrix.}
}
\details{
  This function calls the C function  \code{R_ExpectationCovarianceStatistic}, a
  simple wrapper for the C function
  \code{libcoin_R_ExpectationCovarianceStatistic} defined in the \pkg{libcoin}
  package.

  For more information on this approach, consult section 5.4.2 of the Writing R
  Extensions manual.
}
\value{
  A list.
}
\examples{
n <- 100
p <- 4
q <- 2
X <- matrix(runif(p * n), nc = p)
Y <- matrix(runif(q * n), nc = q)
ExpectationCovarianceStatistic(X, Y)
}
